backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
asyncmy
cryptography
pandas
python-multipart
uvicorn[standard]
websockets